Seperating out the ssh manager to reduce dependence on minimega if it is not needed
