+++ src/firewheel/cli/helpers/ssh
@@ -45,7 +45,7 @@ RUN LocalPython ON control
 
 import sys
 
-from firewheel.cli.utils import SSHManager
+from firewheel.cli.ssh_manager import SSHManager
 
 if __name__ == "__main__":
     # Parse the command line inputs as an SSH command
